Refactor and optimize the design
