+++ dartagnan/src/main/java/com/dat3m/dartagnan/verification/model/event/RegReaderModel.java
@@ -1,10 +1,13 @@
 package com.dat3m.dartagnan.verification.model.event;
 
+import com.dat3m.dartagnan.program.event.RegReader;
 import com.dat3m.dartagnan.program.Register.Read;
 
 import java.util.Set;
 
 
 public interface RegReaderModel extends EventModel {
-    Set<Read> getRegisterReads();
+    default Set<Read> getRegisterReads() {
+        return ((RegReader) this.getEvent()).getRegisterReads();
+    };
 }
