+++ dartagnan/src/main/java/com/dat3m/dartagnan/verification/model/event/RegWriterModel.java
@@ -1,8 +1,11 @@
 package com.dat3m.dartagnan.verification.model.event;
 
+import com.dat3m.dartagnan.program.event.RegWriter;
 import com.dat3m.dartagnan.program.Register;
 
 
 public interface RegWriterModel extends EventModel {
-    Register getResultRegister();
+    default Register getResultRegister() {
+        return ((RegWriter) this.getEvent()).getResultRegister();
+    }
 }
